Avance 3: Integración MySQL + Spring Data JPA
Adición de base de datos a la página

+++ app-backend/focuss-backend/src/main/java/com/focus/focuss_backend/service/TareaService.java
@@ -29,10 +29,14 @@ public class TareaService {
     }
 
     public Tarea update(Tarea tarea) {
-        return tareaRepository.update(tarea);
+        return tareaRepository.save(tarea);
     }
 
     public boolean deleteById(Long id) {
-        return tareaRepository.deleteById(id);
+        if (tareaRepository.existsById(id)) {
+            tareaRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
